First basic Amazon Price Scrape
Right now I established a connection to amazon based off a book title. I also found a basic way to scrape the price of that book

However, right now, the implementation is only working for one hard coded book

Also, there's possiblility for the amazon book to be $0.0 if it's the kindle price. I need to make some way to filter out $0.0 priced books on amazon